1_microservices/5_img-worker/pyproject.toml

[project]
name = "gpu-worker"
version = "0.1.0"
description = "Synthetic GPU worker for ai-microgen"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pillow",
]
